Moved data & constants to utils
